+++ src/test/java/racingcar/CarsTest.java
@@ -2,7 +2,7 @@ package racingcar;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.params.provider.Arguments.arguments;
-import static racingcar.Car.UNIT;
+import static racingcar.domain.Car.UNIT;
 
 import java.util.Arrays;
 import java.util.List;
@@ -16,6 +16,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import racingcar.domain.Car;
+import racingcar.domain.Cars;
 
 class CarsTest {
